Truckenbrodt 99 Buttons
Two buttons were added to both the spot script and the html interface. The first button generates js coded trees for the TohonoOdham portion of Truckenbrodt 99. The other button selects all constraints needed for Truckenbrodt's theory.

Truck99(test) is the file which the first button to generate trees from pulls the js trees.

The Truckenbrodt (1999) Trees file contains all trees from the Truckenbrodt paper, not yet formatted to be pulled.

Also made the button for the Danish tree generation visible for use.